+++ leia-common/src/main/java/com/grookage/leia/common/utils/Utils.java
@@ -1,0 +1,56 @@
+/*
+ * Copyright (c) 2024. Koushik R <rkoushik.14@gmail.com>.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.grookage.leia.common.utils;
+
+import com.grookage.leia.common.exception.SchemaValidationException;
+import com.grookage.leia.common.exception.ValidationErrorCode;
+import lombok.experimental.UtilityClass;
+
+import java.lang.reflect.Field;
+import java.lang.reflect.ParameterizedType;
+import java.lang.reflect.Type;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+@UtilityClass
+public class Utils {
+    public List<Field> getAllFields(final Class<?> type) {
+        List<Field> fields = new ArrayList<>();
+        for (Class<?> c = type; c != null; c = c.getSuperclass()) {
+            fields.addAll(Arrays.asList(c.getDeclaredFields()));
+        }
+        return fields;
+    }
+
+    public Type[] getTypeArguments(final ParameterizedType parameterizedType) {
+        final var typeArguments = parameterizedType.getActualTypeArguments();
+        if (typeArguments.length == 0) {
+            throw SchemaValidationException.error(ValidationErrorCode.INVALID_SCHEMAS,
+                    String.format("No type arguments found for %s", parameterizedType));
+        }
+        return typeArguments;
+    }
+
+    public Set<String> getEnumValues(final Class<?> klass) {
+        return Arrays.stream(klass.getEnumConstants())
+                .map(enumConstant -> ((Enum<?>) enumConstant).name())
+                .collect(Collectors.toSet());
+    }
+}